requirements.txt
django~=3.2.7
psycopg2~=2.9.1
djangorestframework~=3.12.4
coverage~=6.0.1
django-filter~=2.3.0
social-auth-app-django~=5.0.0